requirements.txt
fastapi
uvicorn
numpy
faiss-cpu
Pillow
opencv-python
insightface
onnxruntime